[STATE RESTORE SERIES] Set state when in the middle of a round
This is the first in a series of changes that will allow state to
be set from a serialized state captured from somewhere. As of this
change if the captured state represents play in the middle of a
round, the state can be restored successfully. This is picked as the
starting point for this work because being in the middle of a round
means we skip a number of edge cases where various things might be
empty or uninitialized.