Establish state synchronizionation worker
Add GameStreamer which polls for state updates and pushes new state
updates on demand. Care is taken only to inflate json pulled from the
server if it differs from the current local state or from a state
that this client has pushed. This is done by taking a SHA1 hash of
all outgoing JSON representing states and a SHA1 of the last state to
be pulled from the server. History of pushed states is dropped as that
history is observed from subsequent polls from the server.

Currently this is managed by the GameEngine and GameEngine only
pushes an update on the "trump set pass event". The GameEngine should
push updates for all events that change state. Also we should look
at having something else observe local and remote state changes and
move this logic out of GameEngine.