Add in URL shortener
Use Google URL Shortener to create short link to game state. This
approaches uses the 'browser style' key, which is to say there is
not authentication at all back to our developer ID. Attempts to
authenticate with an 'Android key' where package name and cert
fingerprint would be used have failed. In theory this should be
easy with Play Services, but package compilation fails when trying
to use Play Services 8.4, later research should look into why this
is. In any case, using the browser style key should be sufficient
for now.