Introduce PlayerDisplaysPresenter
The PlayerDisplaysPresenter captures logic about how to handle
when players are seated versus not and what to do when multiple
players are seated ("hot seat mode") versus having a single player
seated.

This change also has a substantial rework of logic related to how
the PlayerDisplay decides which controls are gone vs visible and
enabled vs disabled.