On phones label the card table, other fixes
On phones it can be more difficult to understand who is seated
where since the hands are not arrayed around the table, so on phones
display names around the table.

Also fix small bug where we didn't update team names properly.

Remove save/load buttons.

Fix issue where the name change dialog would contain "DEALER" or
"MAKER" in addition to the player name if that player was that role.

Also, declaring this version 0.1!